civiclens-backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# SQLAlchemy 2.0 & Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0
pyotp==2.9.0  # 2FA TOTP
qrcode==7.4.2  # QR code generation for 2FA setup
bcrypt==4.1.2  # Direct bcrypt (already used)

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/ML
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# File Storage
minio==7.2.3
python-magic==0.4.27
Pillow==10.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
faker==22.4.0

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==1.40.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Testing - Additional
aiosqlite==0.19.0

# File Upload & Processing
python-magic==0.4.27
aiofiles==23.2.1
reportlab==4.0.9
